chore: move s3 service to utility.service package
